Fixed a few things.
* XOR checksum now returns a 16 bit number so that I don't have to shift it
when using in other functions.
* net_bcast_packet() now works properly.

+++ cksum/cksum.h
@@ -29,7 +29,7 @@
     @param  length : Length of data.
     @return XOR sum of data.
 */
-uint8_t xor_sum(net_packet_t *p);
+uint16_t xor_sum(net_packet_t *p);
 
 /**
     @brief  Validate checksum.
